According to ChatGPT moving to centralized package management is the best option.
In the Web.Extensions csproj, the umbraco-extensions template uses a wildcard version, chatgpt says this about it not working in GH actions:

Version="*" is no longer accepted in the restore/pack path your GitHub workflow is using. That wildcard used to limp along in some setups, but with newer SDKs / NuGet (and especially when packing), it’s a hard fail.

Also removed some unnecessary hard references that shouldn't be needed and fixed the regex in the